Develop

See merge request wangxiao_xueyanandroid_business/livevideo!64

+++ src/main/java/com/xueersi/parentsmeeting/modules/livevideo/business/courseware/CoursewarePreload.java
@@ -41,7 +41,6 @@ import java.util.LinkedList;
 import java.util.List;
 import java.util.Locale;
 import java.util.concurrent.CopyOnWriteArrayList;
-import java.util.concurrent.Executor;
 import java.util.concurrent.LinkedBlockingQueue;
 import java.util.concurrent.ThreadPoolExecutor;
 import java.util.concurrent.TimeUnit;
@@ -55,6 +54,7 @@ import ren.yale.android.cachewebviewlib.utils.MD5Utils;
  * Created by: WangDe on 2019/2/27
  */
 public class CoursewarePreload {
+
 
     String TAG = getClass().getSimpleName();
     Logger logger = LoggerFactory.getLogger(TAG);
@@ -91,7 +91,7 @@ public class CoursewarePreload {
     }
 
     List<CoursewareInfoEntity> courseWareInfos = new CopyOnWriteArrayList<>();
-
+    /** 是否紧急下载 */
     AtomicBoolean isPrecise = new AtomicBoolean(false);
 
     AtomicInteger ipPos, cdnPos, ipLength, cdnLength;
@@ -99,7 +99,7 @@ public class CoursewarePreload {
     /**
      * 删除旧的Dir
      */
-    private void deleteOldDir(final File file, final String today) {
+    private synchronized void deleteOldDir(final File file, final String today) {
 //        LiveThreadPoolExecutor executor = LiveThreadPoolExecutor.getInstance();
         executos.execute(new Runnable() {
             @Override
@@ -108,34 +108,49 @@ public class CoursewarePreload {
                 if (file == null || file.listFiles() == null) {
                     return;
                 }
-                for (File itemFile : file.listFiles()) {
-                    if (isCoursewareDir(itemFile.getName()) && !itemFile.getName().equals(today)) {
-                        if (!itemFile.isDirectory()) {
-                            itemFile.delete();
-                        } else {
-                            deleteFor(itemFile);
-                            itemFile.delete();
+                //buglys上面有报Attempt to get length of null array,加上try,catch
+                try {
+                    File[] files = file.listFiles();
+                    if (files == null) return;
+                    for (File itemFile : files) {
+                        //文件夹是日期格式并且不是今天才删除
+                        if (isCoursewareDir(itemFile.getName()) && !itemFile.getName().equals(today)) {
+                            if (!itemFile.isDirectory()) {
+                                itemFile.delete();
+                            } else {
+                                deleteFor(itemFile);
+                                itemFile.delete();
+                            }
                         }
                     }
+                    logger.i("delete file success");
+                    StableLogHashMap hashMap = new StableLogHashMap();
+                    hashMap.put("logtype", " deleteCourseware");
+                    hashMap.put("dir", file.getAbsolutePath());
+                    hashMap.put("sno", "5");
+                    hashMap.put("status", "true");
+                    hashMap.put("ip", IpAddressUtil.USER_IP);
+                    UmsAgentManager.umsAgentDebug(ContextManager.getContext(), UmsConstants.LIVE_APP_ID, LogConfig.PRE_LOAD_START, hashMap.getData());
+                } catch (Exception e) {
+                    logger.e(e);
                 }
-                logger.i("delete file success");
-                StableLogHashMap hashMap = new StableLogHashMap();
-                hashMap.put("logtype", " deleteCourseware");
-                hashMap.put("dir", file.getAbsolutePath());
-                hashMap.put("sno", "5");
-                hashMap.put("status", "true");
-                hashMap.put("ip", IpAddressUtil.USER_IP);
-                UmsAgentManager.umsAgentDebug(ContextManager.getContext(), UmsConstants.LIVE_APP_ID, LogConfig.PRE_LOAD_START, hashMap.getData());
-
             }
         });
     }
 
-    private void deleteFor(final File file) {
+    /**
+     * 循环删除子文件夹
+     *
+     * @param file
+     */
+    private synchronized static void deleteFor(final File file) {
         if (file == null || file.listFiles() == null) {
             return;
         }
-        for (File itemFile : file.listFiles()) {
+
+        File[] files = file.listFiles();
+        if (files == null) return;
+        for (File itemFile : files) {
             if (!itemFile.isDirectory()) {
                 itemFile.delete();
             } else {
@@ -146,7 +161,7 @@ public class CoursewarePreload {
     }
 
     /**
-     * 是否是课件的文件夹
+     * 是否是课件的文件夹(课件文件夹由日期构成)
      *
      * @return
      */
@@ -164,6 +179,7 @@ public class CoursewarePreload {
      * 获取课件信息
      */
     public void getCoursewareInfo(String liveId) {
+        executos.allowCoreThreadTimeOut(true);
         SimpleDateFormat dateFormat = new SimpleDateFormat("yyyyMMdd", Locale.getDefault());
         Date date = new Date();
         final String today = dateFormat.format(date);
@@ -616,9 +632,8 @@ public class CoursewarePreload {
 
 //    private ZipExtractorTask zipExtractorTask;
 
-    // TODO 没有释放核心线程
-    Executor executos = new ThreadPoolExecutor(1, 1,
-            0L, TimeUnit.MILLISECONDS, new LinkedBlockingQueue<Runnable>());
+    ThreadPoolExecutor executos = new ThreadPoolExecutor(1, 1,
+            10L, TimeUnit.SECONDS, new LinkedBlockingQueue<Runnable>());
 
     class ZipDownloadListener implements DownloadListener {
